Added database. Updated backend structure
