+++ src/.env
@@ -1,0 +1,5 @@
+BACKEND_PORT=9091
+ENV=dev
+POSTGRES_PORT=5432
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=p@ssw0rd
