--- src/backend/.env
@@ -1,2 +1,0 @@
-BACKEND_PORT=9091
-ENV=dev
