+++ src/backend/Dockerfile
@@ -2,7 +2,7 @@ FROM node:16-alpine
 WORKDIR /usr/local/apps/backend/ 
 
 COPY . ./
-RUN npm install && npm cache clean --force
+RUN npm install --dev && npm cache clean --force
 RUN npm install --global nodemon 
 
 EXPOSE ${PORT}
