+++ src/backend/package.json
@@ -2,14 +2,14 @@
   "name": "guess-a-song",
   "version": "1.0.0",
   "description": "University project",
-  "main": "index.js",
+  "main": "dist/server.js",
   "directories": {
     "doc": "docs"
   },
   "scripts": {
     "build": "tsc",
-    "start:prd": "node dist/index.js",
-    "start:dev": "nodemon --legacy-watch src/index.ts"
+    "start:prd": "node dist/server.js",
+    "start:dev": "nodemon --legacy-watch src/server.ts"
   },
   "repository": {
     "type": "git",
@@ -22,13 +22,18 @@
   },
   "homepage": "https://github.com/4math/guess-a-song#readme",
   "dependencies": {
+    "body-parser": "^1.20.1",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "pg": "^8.8.0"
+    "pg": "^8.8.0",
+    "sequelize": "^6.25.2"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.12",
     "@types/express": "^4.17.14",
     "@types/node": "^18.11.0",
+    "@types/pg": "^8.6.5",
     "ts-node": "^10.9.1"
   }
 }
