+++ src/backend/src/db/db.config.ts
@@ -1,0 +1,1 @@
+export const pgConnection = {};
