+++ src/docker-compose.yml
@@ -1,5 +1,20 @@
 version: '3.4'
 services:
+  db:
+    container_name: postgresql
+    image: postgres
+    environment:
+      POSTGRES_DB: project
+      POSTGRES_USER: ${POSTGRES_USER:-postgres}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-p@ssw0rd}
+      PGDATA: /data/postgres
+    volumes:
+      - postgres:/data/postgres
+    ports:
+      - 5432:${POSTGRES_PORT}
+    env_file:
+      - .env
+    restart: unless-stopped
   backend:
     container_name: backend
     restart: unless-stopped
@@ -11,11 +26,12 @@ services:
     ports:
       - 9091:${BACKEND_PORT}
     env_file:
-      - ./backend/.env
+      - .env
     volumes:
       - ./backend:/usr/local/apps/backend
-      # - /usr/local/apps/backend/node_modules
       - exclude:/usr/local/apps/backend/node_modules
-
+    depends_on:
+      - db  
 volumes:
   exclude:
+  postgres:
